Just get rid of any Vocab struct for string-to-int mapping altogether
The way Vocab was used doesn't really fit this problem, where all words
have to be generated via segmentation on the fly instead of being something
already known. The constant Dict lookup in the process of converting between
char sequence and words was causing ridiculous overheads.

I may introduce Vocab for only chars later. Or I may find a more efficient
implementation. Or maybe not just do without it in this project.